# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "golang crud api",
        "title": "gocrud",
        "contact": {},
        "version": "1.0"
    },
    "host": "{host}:{port}",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Log in an existing user with the provided email and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "User login request object",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UserLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully logged in",
                        "schema": {
                            "$ref": "#/definitions/domain.UserResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/register": {
            "post": {
                "description": "Register a new user with the specified name, email, and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User registration request object",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UserRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully registered a new user",
                        "schema": {
                            "$ref": "#/definitions/domain.UserResponse"
                        }
                    }
                }
            }
        },
        "/notes": {
            "get": {
                "description": "Retrieve all available notes",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "note"
                ],
                "summary": "Get all notes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter notes by title",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter notes by author",
                        "name": "author",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter notes by user ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter notes by visibility",
                        "name": "visibility",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved all notes",
                        "schema": {
                            "$ref": "#/definitions/domain.NotePaginationResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new note with the specified title, content, and visibility",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "note"
                ],
                "summary": "Create a new note",
                "parameters": [
                    {
                        "description": "Note request object",
                        "name": "note",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.NoteRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created a new note",
                        "schema": {
                            "$ref": "#/definitions/domain.NoteResponse"
                        }
                    }
                }
            }
        },
        "/notes/{id}": {
            "get": {
                "description": "Retrieve a note based on the provided ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "note"
                ],
                "summary": "Get a note by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Note ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved a note by ID",
                        "schema": {
                            "$ref": "#/definitions/domain.NoteResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing note based on the provided ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "note"
                ],
                "summary": "Update a note by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Note ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated note object",
                        "name": "note",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.NoteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated a note by ID",
                        "schema": {
                            "$ref": "#/definitions/domain.NoteResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an existing note based on the provided ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "note"
                ],
                "summary": "Delete a note by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Note ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted a note by ID",
                        "schema": {
                            "$ref": "#/definitions/domain.SuccessResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.Claims": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "domain.Metadata": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "order": {
                    "type": "string",
                    "enum": [
                        "asc",
                        "desc"
                    ]
                },
                "page": {
                    "type": "integer"
                },
                "sort": {
                    "type": "string",
                    "enum": [
                        "id",
                        "user_id",
                        "name",
                        "title",
                        "created_at",
                        "updated_at"
                    ]
                },
                "total_page": {
                    "type": "integer"
                },
                "total_records": {
                    "type": "integer"
                }
            }
        },
        "domain.NoteAuthor": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "domain.NotePaginationResponse": {
            "type": "object",
            "properties": {
                "metadata": {
                    "$ref": "#/definitions/domain.Metadata"
                },
                "notes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.NoteResponse"
                    }
                }
            }
        },
        "domain.NoteRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string",
                    "maxLength": 30
                },
                "user_id": {
                    "type": "integer"
                },
                "visibility": {
                    "type": "string",
                    "enum": [
                        "private",
                        "public"
                    ]
                }
            }
        },
        "domain.NoteResponse": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/domain.NoteAuthor"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "visibility": {
                    "type": "string"
                }
            }
        },
        "domain.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "domain.UserLoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "domain.UserRegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 4
                },
                "password": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 8
                }
            }
        },
...
        "domain.UserResponse": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "tokens": {
                    "$ref": "#/definitions/domain.UserToken"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.UserToken": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "claims": {
                    "$ref": "#/definitions/domain.Claims"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
...
    }
}